apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
aiofiles==23.2.1
Pillow==10.3.0
numpy==1.26.4
scikit-learn==1.4.2
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
prisma==0.13.1
anyio==4.3.0

# Uncomment for real face detection:
# deepface==0.0.93
# tensorflow==2.16.1

# Uncomment for OpenAI classification:
# openai==1.30.0
